% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_distort.R
\name{add_distort}
\alias{add_distort}
\title{add_distort}
\usage{
add_distort(old_tissue_region, sample_angle = pi/6, tissue_angle = 0,
  sigma = 0.01, scaleloc = FALSE, thres = 1)
}
\arguments{
\item{old_tissue_region}{An sf object of the tissue type region polygons.}

\item{sample_angle}{a number of the angles for the whole sample.}

\item{tissue_angle}{a number or a vector of the angles for each of the tissue regions. 
Tissue type regions are rotated clockwise by the input angle.}

\item{sigma}{standard deviation of the normal error}

\item{scaleloc}{whether to scale the location coordinates to [0,1]}

\item{thres}{threshold parameter of the concave hull calculation method. Larger value results in simpler shapes}
}
\value{
An sf object of the polygons of the updated tissue type regions.
}
\description{
The function adds distortions to the input tissue type regions.
}
\details{
The function adds distortions to the input tissue type regions by adding 
normal errors to the coordinates of the vertices of the tissue type region polygons.
}
\references{
Pebesma, E., 2018. Simple Features for R: Standardized Support for 
Spatial Vector Data. The R Journal 10 (1), 439-446, 
https://doi.org/10.32614/RJ-2018-009
}
