% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cell_types.R
\name{assign_cell_types}
\alias{assign_cell_types}
\title{assign_cell_types}
\usage{
assign_cell_types(ttype, nttype, cell_type_proportion = NULL)
}
\arguments{
\item{ttype}{tissue types of the cells}

\item{nttype}{total number of possible tissue types. should be no less than the max value of ttype and the 
number of unique values in ttype. some tissue types may not exist in ttype}

\item{cell_type_proportion}{matrix of cell type proportion, cell type (row) by tissue type (column)}
}
\value{
A vector of the assigned cell type
}
\description{
The function chooses the cell type for each cell.
}
\details{
This function choose the cell type for each cell. Each tissue type could contains
multiple types of cells. Based on the cell type proportion, this function randomly samples 
the cell type for each cell based on their corresponding tissue type.
}
