% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_tissue_types.R
\name{assign_tissue_types}
\alias{assign_tissue_types}
\title{assign_tissue_types}
\usage{
assign_tissue_types(tissue_region, cell_spot_position, nneighbor = 3)
}
\arguments{
\item{tissue_region}{sf object of tissue region polygons.}

\item{cell_spot_position}{coordinates for the cells}

\item{nneighbor}{number of nearest tissue region taken into account}
}
\value{
A vector of the tissue region type
}
\description{
The function assigns to each cell the corresponding tissue type.
}
\details{
This function assigns the cells to a tissue type 
if the cell is located in the tissue region polygon
}
\references{
Pebesma, E., 2018. Simple Features for R: Standardized Support for 
Spatial Vector Data. The R Journal 10 (1), 439-446, 
https://doi.org/10.32614/RJ-2018-009
}
