% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_position_simulation.R
\name{cell_position_simulation}
\alias{cell_position_simulation}
\title{cell_position_simulation}
\usage{
cell_position_simulation(sample_vertices, ncells = 10000)
}
\arguments{
\item{sample_vertices}{coordinates for the vertices of sample polygon. Each row for a vertex.}

\item{ncells}{number of cells to simulate}
}
\value{
A matrix of the centers, 1st column as x coordinate and 2nd column for the y coordinate
}
\description{
The function simulates the coordinates of the cells within the sample region.
}
\details{
This function simulates the location of the cells within the sample region
}
\references{
Pebesma, E., 2018. Simple Features for R: Standardized Support for 
Spatial Vector Data. The R Journal 10 (1), 439-446, 
https://doi.org/10.32614/RJ-2018-009
}
