% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{changeLibSize}
\alias{changeLibSize}
\title{changeLibSize
This function modifies the simulated counts so that the total 
read count in each spot is the same as the example data.}
\usage{
changeLibSize(count, nreadspot, weight)
}
\arguments{
\item{count}{simulated matrix of read count, rows as genes and columns as spots}

\item{nreadspot}{number of total read count in each spot in the example data}

\item{weight}{a matrix of probability weights for the read count of each gene in each cell being changed.}
}
\value{
a matrix of the updated read count
}
\description{
changeLibSize
This function modifies the simulated counts so that the total 
read count in each spot is the same as the example data.
}
\author{
Qi Gao
}
