% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cumsum_cpp}
\alias{cumsum_cpp}
\title{cumsum_cpp
This function calculates the cumulative sum (partial sum) of a vector.}
\usage{
cumsum_cpp(x)
}
\arguments{
\item{x}{vector to be summed.}
}
\value{
vector of the cumulative sum (partial sum)
}
\description{
cumsum_cpp
This function calculates the cumulative sum (partial sum) of a vector.
}
\author{
Qi Gao
}
