% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{downSampleRead}
\alias{downSampleRead}
\title{downSampleRead
This function uses down-sampling to obtain the reads in each spot.}
\usage{
downSampleRead(count, nread)
}
\arguments{
\item{count}{gene (row) by cell (column) matrix. Gene read count in each cell.}

\item{nread}{cell (row) by spot (column) matrix. Target number of reads sampled from each cell in each spot.}
}
\value{
gene (row) by spot (column) matrix. Gene expression level in each spot.
}
\description{
downSampleRead
This function uses down-sampling to obtain the reads in each spot.
}
\author{
Qi Gao
}
