% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_bound.R
\name{est_bound}
\alias{est_bound}
\title{est_bound}
\usage{
est_bound(sploc, ttype, nttype, thres = 1)
}
\arguments{
\item{sploc}{coordinates for the cells.}

\item{ttype}{tissue types of the cells. should be integers starting from 1.}

\item{nttype}{total number of possible tissue types. should be no less than the max value of ttype and the 
number of unique values in ttype. some tissue types may not exist in ttype}

\item{thres}{threshold parameter of the concave hull calculation method. Larger value results in simpler shapes}
}
\value{
A sf object of the polygons of the tissue type regions estimated from data.
}
\description{
The function estimates tissue type regions using cell/spot locations and the corresponding tissue type.
}
\details{
The function estimates tissue type regions using cell/spot locations and the corresponding tissue type
from reference data.
}
\references{
Joel Gombin and Ramnath Vaidyanathan and Vladimir Agafonkin (2020). 
concaveman: A Very Fast 2D Concave Hull Algorithm. R package 
version 1.1.0. https://CRAN.R-project.org/package=concaveman

Pebesma, E., 2018. Simple Features for R: Standardized Support for 
Spatial Vector Data. The R Journal 10 (1), 439-446, 
https://doi.org/10.32614/RJ-2018-009
}
