% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_gamma}
\alias{estimate_gamma}
\title{estimate_gamma
This function estimates gamma distribution shape and rate parameters.}
\usage{
estimate_gamma(x)
}
\arguments{
\item{x}{target data vector}
}
\value{
list of gamma distribution shape and rate parameters
}
\description{
estimate_gamma
This function estimates gamma distribution shape and rate parameters.
}
\author{
Qi Gao
}
