% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{estimate_lognormal}
\alias{estimate_lognormal}
\title{estimate_lognormal
This function estimates estimate_lognormal distribution loc and scale parameters.}
\usage{
estimate_lognormal(x)
}
\arguments{
\item{x}{target data vector}
}
\value{
list of lognormal distribution loc and scale parameters
}
\description{
estimate_lognormal
This function estimates estimate_lognormal distribution loc and scale parameters.
}
\author{
Qi Gao
}
