% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nb.R
\name{estimate_nb}
\alias{estimate_nb}
\title{estimate_nb
This function estimates negative binomial distribution mean and dispersion parameters.}
\usage{
estimate_nb(datavec)
}
\arguments{
\item{datavec}{target data vector}
}
\value{
vector of negative binomial distribution mean (1st element), dispersion (2nd element) parameters and 0 (3rd element).
}
\description{
estimate_nb
This function estimates negative binomial distribution mean and dispersion parameters.
}
\author{
Qi Gao
}
