% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nb.R
\name{estimate_zinb}
\alias{estimate_zinb}
\title{estimate_zinb
This function estimates the mean, dispersion and zero inflation parameters in zero-inflated negative binomial distribution.}
\usage{
estimate_zinb(data, maxiter = 400)
}
\arguments{
\item{data}{target data matrix, rows as genes and columns as spots}

\item{maxiter}{maximum number of iteration for R optim}
}
\value{
matrix of negative binomial distribution mean (1st column), dispersion (2nd column) and zero-inflation (3rd column) parameters.
}
\description{
estimate_zinb
This function estimates the mean, dispersion and zero inflation parameters in zero-inflated negative binomial distribution.
}
\author{
Qi Gao
}
