% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScape.R
\name{GeneScape_est}
\alias{GeneScape_est}
\title{GeneScape_est}
\usage{
GeneScape_est(data, groups = NULL)
}
\arguments{
\item{data}{target gene (row) by cell (column) data matrix}

\item{groups}{group of the cells}
}
\value{
list of two gene (row) by group (column) matrices. first matrix contains 
the mean parameter, while second matrix contains zero inflation parameters.
}
\description{
This function estimate zero-inflated poisson distribution parameters from data
}
