% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScape.R
\name{GeneScape_par}
\alias{GeneScape_par}
\title{GeneScape_par}
\usage{
GeneScape_par(nCells = 6000, nGroups = NULL, groups = NULL,
  lib.size = NULL, lib.size.mean = 10000, lib.size.sd = 2000,
  de.fc.mat = NULL, nGenes = 5000, gene.mean.shape = 0.3,
  gene.mean.rate = 0.15, gene.means = NULL, zero.inflat = NULL,
  add.zero.inflat = TRUE, zero.inflat.times = 0.5, de.n = 50,
  de.share = NULL, de.id = NULL, de.fc.mean = 1, de.fc.sd = 0.2,
  add.sub = FALSE, sub.major = NULL, sub.prop = 0.1, sub.group = NULL,
  sub.de.n = 20, sub.de.id = NULL, sub.de.common = FALSE,
  sub.de.fc.mean = 1, sub.de.fc.sd = 0.2)
}
\arguments{
\item{nCells}{number of cells}

\item{nGroups}{number of cell groups}

\item{groups}{group information for cells}

\item{lib.size}{library size for cells}

\item{lib.size.mean}{location parameter for library size (log-normal distribution)}

\item{lib.size.sd}{scale parameter for library size (log-normal distribution)}

\item{de.fc.mat}{differential expression fold change matrix, could be generated by this function}

\item{nGenes}{number of genes}

\item{gene.mean.shape}{shape parameter for mean expression level (Gamma distribution)}

\item{gene.mean.rate}{rate parameter for mean expression level (Gamma distribution)}

\item{gene.means}{mean gene expression levels}

\item{zero.inflat}{vector of zero inflation paramters}

\item{add.zero.inflat}{whether to add zero inflation}

\item{zero.inflat.times}{the zero inflation paramter is simulated as zero.inflat.times / (1 + exp(gene.means))}

\item{de.n}{number of differentially expressed genes in each cell type. Should be a integer or a vector of length nGroups}

\item{de.share}{number of shared DE genes between neighbor cell types. Should be a vector of length (nGroups - 1)}

\item{de.id}{the index of genes that are DE across cell types. Should be a list of vectors. 
Each vector corresponds to a cell type. With non-null value of de.id, de.n and de.share would be ignored.}

\item{de.fc.mean}{the location parameter for the fold change of DE genes. Should be a number, a vector of length nGroups}

\item{de.fc.sd}{the scale parameter for fold change (log-normal distribution). Should be a number or a vector of length nGroups}

\item{add.sub}{whether to add sub-cell-types}

\item{sub.major}{the major cell types correspond to the sub-cell-types}

\item{sub.prop}{proportion of sub-cell-types in the corresponding major cell type}

\item{sub.group}{cell index for sub-cell-types. With non-null sub.group specified, sub.prop would be ignored.}

\item{sub.de.n}{number of differentially expressed genes in each sub-cell-type compared to the corresponding major cell type. Should be a integer or a vector of length sub.major}

\item{sub.de.id}{the index of additional differentially expressed genes between sub-cell-types and the corresponding major cell types}

\item{sub.de.common}{whether the additional differential expression structure should be same for all sub-cell-types}

\item{sub.de.fc.mean}{similar to de.fc.mean, but for addtional differentially expressed genes in sub-cell-types}

\item{sub.de.fc.sd}{similar to de.fc.sd, but for addtional differentially expressed genes in sub-cell-types}
}
\value{
A list of simulation parameters: vector of cell groups, vector of cell library size, gene expression 
parameter (zero-inflated poisson model), and matrix of differentially expression fold changes.
}
\description{
This function simulate parameters for single cell RNAseq data simulation with complicated differential expression.
}
\examples{
set.seed(1)
para <- GeneScape_par()
}
\references{
Zappia, L., Phipson, B., & Oshlack, A. (2017). Splatter: Simulation of single-cell RNA sequencing data. Genome Biology, 18(1). https://doi.org/10.1186/s13059-017-1305-0
}
