% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScape_sim.R
\name{GeneScape_sim}
\alias{GeneScape_sim}
\title{GeneScape_sim
This function generates simulated data based on an example data.}
\usage{
GeneScape_sim(data, ttype, newttype, rank)
}
\arguments{
\item{data}{Example data matrix}

\item{ttype}{Original tissue type of each spot}

\item{newttype}{output tissue type of each spot}

\item{rank}{whether to rank the expression level in each tissue type based on the example}
}
\value{
matrix of simulated data
}
\description{
GeneScape_sim
This function generates simulated data based on an example data.
}
\author{
Qi Gao
}
