% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScapeS_sim.R
\name{GeneScapeS_est}
\alias{GeneScapeS_est}
\title{GeneScapeS_est
This function estimates distribution parameters from an example data.}
\usage{
GeneScapeS_est(data, ttype)
}
\arguments{
\item{data}{Example data matrix}

\item{ttype}{Original tissue type of each spot}
}
\value{
list of matrices of negative binomial distribution mean (1st column), 
dispersion (2nd column) and zero-inflation (3rd column) parameters. one matrix for each tissue type.
}
\description{
GeneScapeS_est
This function estimates distribution parameters from an example data.
}
\author{
Qi Gao
}
