% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScapeS_sim.R
\name{GeneScapeS_sim}
\alias{GeneScapeS_sim}
\title{GeneScapeS_sim
This function generates simulated data based on distribution parameters estimated from an example data.}
\usage{
GeneScapeS_sim(para, newttype, rank = TRUE, data = NULL, ttype = NULL)
}
\arguments{
\item{para}{Example data matrix}

\item{newttype}{output tissue type of each spot}

\item{rank}{whether to rank the expression level in each tissue type based on the example}

\item{data}{original reference dataset used for create para}

\item{ttype}{Original tissue type of each spot}
}
\value{
matrix of simulated data
}
\description{
GeneScapeS_sim
This function generates simulated data based on distribution parameters estimated from an example data.
}
\author{
Qi Gao
}
