% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nb.R
\name{inverse_logit}
\alias{inverse_logit}
\title{inverse_logit
This function is the inverse logit function.}
\usage{
inverse_logit(x)
}
\arguments{
\item{x}{number}
}
\value{
inverse logit function value with input x
}
\description{
inverse_logit
This function is the inverse logit function.
}
\author{
Qi Gao
}
