% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_tissue_update.R
\name{knn_tissue_update}
\alias{knn_tissue_update}
\title{knn_tissue_update}
\usage{
knn_tissue_update(cell_spot_position, old_tissue_type, k)
}
\arguments{
\item{cell_spot_position}{coordinates for the cells}

\item{old_tissue_type}{previously assigned tissue type for each cell. should be integers start from one}

\item{k}{number of nearest neighbor}
}
\value{
A vector of the updated tissue type
}
\description{
The function updates the tissue type for each cell.
}
\details{
This function assigns to the each cell to a new tissue type 
by sampling the tissue types of its k nearest neighbors.
}
