% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{knnassign}
\alias{knnassign}
\title{knnassign
This function uses KNN sampling to simulate gradually changed tissue types.}
\usage{
knnassign(distmat, ttype, nttype, k)
}
\arguments{
\item{distmat}{Spot distance matrix}

\item{ttype}{Input tissue type of each spot}

\item{nttype}{total number of possible tissue types. should be no less than the max value of ttype and the 
number of unique values in ttype. some tissue types may not exist in ttype}

\item{k}{parameter for KNN sampling}
}
\value{
List of tissue type weight and updated tissue type with gradual change
}
\description{
knnassign
This function uses KNN sampling to simulate gradually changed tissue types.
}
\author{
Qi Gao
}
