% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_distort.R
\name{rotate_pg}
\alias{rotate_pg}
\title{rotate_pg}
\usage{
rotate_pg(pg, angle = pi/12, center = NULL)
}
\arguments{
\item{pg}{An sf object of the polygon.}

\item{angle}{a number of the angle. The polygon is rotated clockwise by the input angle.}

\item{center}{customized rotation center}
}
\value{
An sf object of the rotated polygon.
}
\description{
The function rotate polygons clockwise.
}
\details{
The function rotate polygons clockwise based on the input angle around the center (by default its centroid).
}
\references{
Pebesma, E., 2018. Simple Features for R: Standardized Support for 
Spatial Vector Data. The R Journal 10 (1), 439-446, 
https://doi.org/10.32614/RJ-2018-009
}
