% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_distort.R
\name{rotate_sample}
\alias{rotate_sample}
\title{rotate_sample}
\usage{
rotate_sample(tissue_region, angle = pi/12, thres = 1)
}
\arguments{
\item{tissue_region}{Tissue region polygons.}

\item{angle}{a number of the angle. The polygon is rotated clockwise by the input angle.}

\item{thres}{threshold parameter of the concave hull calculation method. Larger value results in simpler shapes}
}
\value{
An sf object of the rotated tissue regions
}
\description{
The function rotate polygons clockwise.
}
\details{
The function rotate tissue regions clockwise based on the input angle around the center (by default its centroid).
}
\references{
Pebesma, E., 2018. Simple Features for R: Standardized Support for 
Spatial Vector Data. The R Journal 10 (1), 439-446, 
https://doi.org/10.32614/RJ-2018-009
}
