% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_int}
\alias{sample_int}
\title{sample_int
This function samples integers from 1 to an upper limit integer value with equal probability without replacement.}
\usage{
sample_int(maxvalue, nsample)
}
\arguments{
\item{maxvalue}{the max integer to be sampled.}

\item{nsample}{the number of integer to be sampled.}
}
\value{
vector of the integers in the sample
}
\description{
sample_int
This function samples integers from 1 to an upper limit integer value with equal probability without replacement.
}
\author{
Qi Gao
}
