% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_region_simulation.R
\name{sample_region_simulation}
\alias{sample_region_simulation}
\title{sample_region_simulation}
\usage{
sample_region_simulation(nvert = 13, sample_shape = c("square", "circle"),
  limitx = c(0, 1), limity = c(0, 1))
}
\arguments{
\item{nvert}{number of vertices of the sample polygon}

\item{sample_shape}{shape of the capture area}

\item{limitx}{min and max value of x coordinates}

\item{limity}{min and max value of y coordinates}
}
\value{
A matrix of the coordinates of sample polygon vertices
}
\description{
This function simulates the sample region.
}
\details{
This function simulates the sample region with the capture area as a polygon.
}
\references{
Valtr, P. Probability that n random points are in convex position. 
Discrete Comput Geom 13, 637-643 (1995). https://doi.org/10.1007/BF02574070
}
