% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_distort.R
\name{scale_sample}
\alias{scale_sample}
\title{scale_sample}
\usage{
scale_sample(tissue_region, scale_prop = c(1, 1), thres = 1)
}
\arguments{
\item{tissue_region}{An sf object of the tissue type region polygons.}

\item{scale_prop}{a vector of length 2, for x and y correspondingly. Value greater than 1
would enlarge the tissue type region on the corresponding coordinates.}

\item{thres}{threshold parameter of the concave hull calculation method. Larger value results in simpler shapes}
}
\value{
An sf object of the polygons of the updated tissue type regions.
}
\description{
The function shifts the tissue type regions.
}
\details{
The function shifts tissue type regions by modifying the x and y 
coordinates of the vertices of the tissue type region polygons. The sizes of modifications
equal to the input proportion times the difference between the maximum of x (y) coordinates 
and the minimum of x (y) coordinates
}
\references{
Pebesma, E., 2018. Simple Features for R: Standardized Support for 
Spatial Vector Data. The R Journal 10 (1), 439-446, 
https://doi.org/10.32614/RJ-2018-009
}
