% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_distort.R
\name{shift_sample}
\alias{shift_sample}
\title{shift_sample}
\usage{
shift_sample(tissue_region, shift_prop = c(0, 0))
}
\arguments{
\item{tissue_region}{An sf object of the tissue type region polygons.}

\item{shift_prop}{a vector of length 2, for x and y correspondingly. Large absolute value
implies greater shift. Positive value implies positive change of x and y coordinates (shift up and right)}
}
\value{
An sf object of the polygons of the updated tissue type regions.
}
\description{
The function scales the tissue type regions.
}
\details{
The function scales the tissue type regions around the sample centroid
}
\references{
Pebesma, E., 2018. Simple Features for R: Standardized Support for 
Spatial Vector Data. The R Journal 10 (1), 439-446, 
https://doi.org/10.32614/RJ-2018-009
}
