% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sp_dist_euclidean_cpp}
\alias{sp_dist_euclidean_cpp}
\title{sp_dist_euclidean_cpp
This function calculates Euclidean distance between cell locations and spot locations.}
\usage{
sp_dist_euclidean_cpp(cloc, sloc)
}
\arguments{
\item{cloc}{matrix of location coordinates of the cells. Each row represents a cell.}

\item{sloc}{matrix of location coordinates of the spots. Each row represents a spot.}
}
\value{
matrix of the Euclidean distance, number of cells (row) by number of spots (column)
}
\description{
sp_dist_euclidean_cpp
This function calculates Euclidean distance between cell locations and spot locations.
}
\author{
Qi Gao
}
