% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot_simulate.R
\name{spot_simulate}
\alias{spot_simulate}
\title{spot_simulate}
\usage{
spot_simulate(dimension = c(78, 64), n = 5000, structure = c("interlaced",
  "uniform", "random"), error_prop = 0.04, limitx = c(0, 1), limity = c(0,
  1))
}
\arguments{
\item{dimension}{dimension of the spots, only used for "interlaced" and "uniform" structure.}

\item{n}{number of spots, only used for "random" structure}

\item{structure}{structure of the spots}

\item{error_prop}{size of error added to the coordinates of the spots}

\item{limitx}{min and max value of x coordinates}

\item{limity}{min and max value of y coordinates}
}
\value{
A matrix of the spot coordinates
}
\description{
The function simulates the spot positions.
}
\details{
This function simulates the spot coordinates. The "interlaced" structure 
mimic the Visium sequencing data. The "uniform" structure generate spot positions
that are located on the vertices of squares. And the "random" structure randomly 
sample the spot locations.
}
