% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_region_simulation.R
\name{tissue_region_simulation}
\alias{tissue_region_simulation}
\title{tissue_region_simulation}
\usage{
tissue_region_simulation(sample_vertices, ntissuetype = 4, centers = NULL)
}
\arguments{
\item{sample_vertices}{coordinates for the vertices of sample polygon. Each row for a vertex.}

\item{ntissuetype}{number of tissue types}

\item{centers}{centers of the tissue types}
}
\value{
A sf object of the polygons of the tissue type regions.
}
\description{
The function simulates regions of different tissue types within the sample region.
}
\details{
This function simulates the tissue type regions using voronoi polygons 
correponding to the simulated or input centers
}
\references{
Pebesma, E., 2018. Simple Features for R: Standardized Support for 
Spatial Vector Data. The R Journal 10 (1), 439-446, 
https://doi.org/10.32614/RJ-2018-009
}
