% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_sample_region.R
\name{update_sample_region}
\alias{update_sample_region}
\title{update_sample_region}
\usage{
update_sample_region(tissue_region, thres = 1)
}
\arguments{
\item{tissue_region}{An sf object of the tissue type region polygons.}

\item{thres}{threshold parameter of the concave hull calculation method. Larger value results in simpler shapes}
}
\value{
An sf object of the polygons of the updated tissue type regions.
}
\description{
The function updates the sample region based on the regions of different tissue types.
}
\details{
The function updates the sample region by finding the concave hull of the vertices of tissue type region polygons.
}
\references{
Pebesma, E., 2018. Simple Features for R: Standardized Support for 
Spatial Vector Data. The R Journal 10 (1), 439-446, 
https://doi.org/10.32614/RJ-2018-009
}
