% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_nb.R
\name{zinb_nll}
\alias{zinb_nll}
\title{zinb_nll
This function calculates the negative log likelihood of zero-inflated negative binomial distribution.}
\usage{
zinb_nll(y, par)
}
\arguments{
\item{y}{observations}

\item{par}{vector of 3 elements. parameters (mean, dispersion, zero-inflation) for zero-inflated negative binomial distribution.}
}
\value{
negative log likelihood of zero-inflated negative binomial distribution
}
\description{
zinb_nll
This function calculates the negative log likelihood of zero-inflated negative binomial distribution.
}
\author{
Qi Gao
}
