% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_expression.R
\name{aggregate_expression}
\alias{aggregate_expression}
\title{aggregate_expression}
\usage{
aggregate_expression(counts, cell_position, spot_position,
  dis_thres_prop = 0.03, sigma_prop = 0.01, lib_size_mean = 5000,
  lib_size_dispersion = 3, cell_annotation = NULL)
}
\arguments{
\item{counts}{read count matrix}

\item{cell_position}{coordinates of the cells}

\item{spot_position}{coordinates of the spots}

\item{dis_thres_prop}{radius of the spot}

\item{sigma_prop}{standard deviation of normal distribution. larger sigma implies that cells 
outside of the spot are more likely to contribute to the read observed in the spot}

\item{lib_size_mean}{mean parameter of target library size (negative binomial distribution)}

\item{lib_size_dispersion}{dispersion parameter of target library size (negative binomial distribution)}

\item{cell_annotation}{cell type annotations}
}
\value{
A matrix of the gene (row) expression in each spot (column)
}
\description{
The function aggregate the gene expression level in high resolution data to generate low resolution data.
}
\details{
This function  aggregate the gene expression level in high resolution (cell level) data to 
generate low resolution (spot level) data. Cells contribute to the spots based on the distance between 
cell centers and spot centers, with the cells inside the spot contribute all their reads to the correponding 
spot. After obtaining the read pool, the funtion also downsample the reads to a target mean spot library size.
}
