% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexSCsimulation.R
\name{fcsim}
\alias{fcsim}
\title{fcsim}
\usage{
fcsim(n.gene, de.id, fc.loc, fc.scale)
}
\description{
This function similate differential expression fold change level
}
\references{
Zappia, L., Phipson, B., & Oshlack, A. (2017). Splatter: Simulation of single-cell RNA sequencing data. Genome Biology, 18(1). https://doi.org/10.1186/s13059-017-1305-0
}
