% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneScape.R
\name{GeneScape_sim}
\alias{GeneScape_sim}
\title{GeneScape_sim}
\usage{
GeneScape_sim(para, add.path = FALSE, path.n = 4, path.size = 20,
  path.cor = 0.7, path.id = NULL, band.width = 10, add.hub = FALSE,
  hub.n = 10, hub.size = 20, hub.cor = 0.4, hub.id = NULL,
  hub.fix = NULL)
}
\arguments{
\item{para}{simulation parameter estimated by GeneScape_est or simulated by GeneScape_par}

\item{add.path}{whether to add pathways (correlated genes)}

\item{path.n}{number of pathways included. Should be a integer.}

\item{path.size}{number of correlated genes (length of pathway). Should be a number or a vector of length path.n}

\item{path.cor}{correlation parameters}

\item{path.id}{gene index of correlated (pathway) genes. Should be a list of vectors, with each vector represents a pathway. With non-null value of path.id, path.n would be ignored.}

\item{band.width}{No correlation exists if distance of 2 genes are further than band_width in a pathway}

\item{add.hub}{whether to add hub genes}

\item{hub.n}{number of hub genes included. Should be a integer.}

\item{hub.size}{number of genes correlated to the hub gene. Should be a number or a vector of length hub.n}

\item{hub.cor}{correlation parameters between hub genes and their correlated genes}

\item{hub.id}{gene index of hub genes. Should be a list of vectors. With non-null value of hub.id, hub.n would be ignored.}

\item{hub.fix}{user defined genes correlated to hub genes (others are randomly selected). Should be a list of vectors of length hub.n or same as hub.id.}
}
\value{
A list of read count data, cell groups, cell library size, gene mean expression, gene differential expression rate, pathway genes and hub gene indices.
}
\description{
This function generate single cell RNAseq data using simulation parameter.
}
\examples{
set.seed(1)
para <- GeneScape_par()
data <- GeneScape_sim(para)
}
\references{
Zappia, L., Phipson, B., & Oshlack, A. (2017). Splatter: Simulation of single-cell RNA sequencing data. Genome Biology, 18(1). https://doi.org/10.1186/s13059-017-1305-0
}
